requirements.txt
python-dotenv>=1.0,<2.0
pytest>=7.0,<9.0
ruff>=0.6,<0.7
httpx>=0.26,<0.28
psycopg[binary]>=3.1,<4.0
fastapi>=0.115,<0.116
uvicorn[standard]>=0.32,<0.33
pydantic>=2.0,<3.0
python-multipart>=0.0.9,<0.1.0
sqlalchemy[asyncio]>=2.0,<3.0
aiosqlite>=0.20,<0.21
asyncpg>=0.29,<0.30
